Support immutable container instance in menus.
This broke a while ago with the Container/Elements refactor. Puzzle pirates never ended up using
scrolling menus, opting for paging instead, but now someone on the forums has pointed out they don't
work (including the one in the demo).

The edge case of mutable non-elements containers is not prevented but warned against.