Option for ignoring unchanged flump libs.
Also, as I was driving by, made it a maven plugin. I think this is the right thing to do and that
no one else is using binary flump libs. But, if it does bother anyone, we could have a library with
the command line interface and the maven plugin could depend on that.

NOTE for IDE users: the artifact id changed, which sometimes means you need to reimport.