Interpolator simplification.
They all effectively operate on a single value (v) between 0 and 1. That value
can be obtained via dt/t when you're driving your interpolation via elapsed
time, and the result can be converted to start + result * range when you want
your result to be in [start, start+range]. But none of those extra bits need to
be in the interp code, obscuring the basic math.