Emit a signal when a Root is validated.
It's needless abuse of inheritance to override a view and override validate()
just to know when a root has been validated. It's also sloppy to just "wait one
frame" and assume the root will be validated after that.

Much better to allow something simple and explicit like:

root.validated().connect(new UnitSlot() {
  public void onEmit () {
    // do whatever we need after the root is validated
  }
}).once();