Reworked ScreenSpace.UIScreen.
Unlike ScreenStack.UIScreen, here we do a little bit for you. We have
createUI() wherein you create one or more roots, and then when the screen
sleeps, we'll dispose those roots for you.

This is feasible here because we don't have wasAdded/wasShown, we just have
wake, so we know you want to create your UI in one place.