Element can now track bindings to reactive values.
This allows the element to establish a connection when it is added to a UI
hierarchy and dispose of it when it is removed. Doing so relieves the caller of
managing the lifetime of the connection themselves, which is error-prone and
leaky.

Often the lifetime of a reactive value and a UI element that displays it do not
coincide, and the reactive value long outlives the element, which results in
leaks if special care is not taken to ensure that connections are closed when
the element goes away. This bind mechanism makes that easy to do, syntactically
concise, and cheaper than creating a Closeable.Set and binding it to the
element's hierarchy signal (which was the only general purpose way to do this
before this change).