A bit of LayoutData revampery, to simplify.
Previously LayoutData had three jobs:

1. Track any resolved style or ephemeral data which is needed to layout a
widget. This stuff is only needed for layout, so we'd resolve it, compute
preferred size, do the layout, then throw it away. That way we're not keeping
around a bunch of crap for every single widget during normal operation.

2. Use the resolved data to compute widget's preferred size, via
LayoutData.computeSize().

3. Use the resolved data to lay the widget out, via LayoutData.layout().

But this sucked when you just wanted to make a minor tweak to computeSize() or
layout() based on some non-ephemeral state. You had to create a special
subclass of LayoutData just to hold the code that really should have been in
the Element subclass instead.

There was also some fiddly shenanigans with BaseLayoutData and LayoutData for
probably stupid reasons that have long since been forgotten.

Now there's just one LayoutData class, which defines two ("NOOP" by default)
methods: computeSize() and layout(), and two Element methods (of the same name)
which delegate to the LayoutData methods by default.

In cases where a simple tweak is needed (Box, Composite, GlyphWidget,
ImageButton), no custom LayoutData class is needed at all. The Element
subclasses just override the methods that would have delegated to LayoutData
and do the right thing right there. Simple!

In one case (Scroller), ScrollerLayoutData is used to track some ephemeral
data, but we still handle the computeSize() and layout() in Scroller's methods
rather than ScrollerLayoutData's methods (mainly because we're specializing
Composite/Container, which now do their layout in Element.layout).

In a few other cases (TextWigdet, Slider), we still do things the "old" way,
where TextWidgetLayoutData and SliderLayoutData implement computeSize() and
layout() directly, because they resolve a shit ton of layout data and it's nice
to be able to access that stuff directly via the this pointer.

I was tempted to convert everything to "Scroller"-style where all LayoutData
subclasses were just structs that held extra data and all custom logic went
into the Element subclass, but I needed to preserve the old LayoutData methods
and behavior for all the custom code out in the wild that has created its own
LayoutData subclass. So since that has to live on, I might as well use it in
the few cases where it's tidier.