Adapt to new ByteBuffer world order.
I was tempted to revamp everything to read from ByteBuffer directly, but the
one stick wicket was DataInputStream.readUTF() which does a bunch of complex
stuff that would be hard to mimic in ByteBuffer-only code. So the first time
someone named a layer with some crazy UTF8 surrogate pair with special
modifying characters or something, shit would break. So nevermind.