Add a reactive sizeValue() to Screen.
This makes it easier to bind the size of a Root to the whole screen and have it
update automatically if the screen size changes due to orientation change.

The value is set up so that listeners to it can be fire and forget as long as
their lifecycle is the same as the screen. When the screen goes away, listeners
to the value all go away too, which would not happen if the sizeValue() was a
transformation of the Graphics.deviceOrient value.