Fix ordering issue in set.
We need to call removeFromParent before we assign the new child as the contents
of the box because *we* might be the old parent.

In most cases this won't ever happen because the public set() checks if we're
adding the same element to the box; but in the transition code (where we fade
or flip a new child into the box), we "pre-add" a child, then transition to
that child during a time when the old and new children are both parented by the
box, and finally we "add" the new child, which causes it to be removed from its
old parent, which in this case is the current box.

This also more closely matches Elements, which calls removeFromParent, then
adds new children to its internal tracking array, then calls didAdd.