Improve our alignment to physical pixel boundaries.
Previously we aligned elements to integral logical pixel boundaries, and relied
on the ratio between logical and physical pixels to be integral. This is no
longer the case in the wild world of zillions of screen sizes and densities.

Our new approach is to leave logical position/size alone until validation, at
which point we ensure that our boundaries (top, left, bottom, right) fall on
integral physical pixels and then translate those back to an updated logical
position and size which are used when laying out deeper children.