Update ModalShadow.java
I noticed that I'm using ModalShadow in many cases as a background layer to a popup that doesn't need interaction (such as an alert dialog) and as such it doesn't have a close button. Users tend to touch/click on the shadow to try and close the popup and if nothing happens they get confused.

Hence this proposal to make ModalShadow optionally emit a _closed_ signal which can then be used to close the shadow and the root used.