Add repeat-part feature background size extends
Add a feature that would allow not only to scale the specified parts of the texture but also to repeat them. This feature can be turned on and off by a corresponding setter. Also add a setter that control wether the repeated parts are aligend at the left or the right side for the horizontal axis and one controlling the bottom and top alignment for the vertical axis.

corrected worng package name

remove unnecessary import

Remove unnecessary reformating code
Adapt formatting of new code

Controll alignment of the repeated parts

Use absolute coordinates for clipping now