SizableGroup: Fix not being resizable at all.
The previous implementation just overrides createLayoutData(float,float) to
return a SizableLayoutData instance but apparently layoutData is not taken
into account neither in:
 - Container.computeSize(LayoutData,float,float) nor in
 - Container.layout(LayoutData,float,float,float,float)
as they both delegate to Layout.

Changed the implementation to mimic the behaviour of the SizableLayoutData
instance created in the previous implementation in createLayoutData,
but using Layout.computeSize as base instead of LayoutData.computeSize.
So we just override computeSize(LayoutData,float,float) to call super
(with adjusted hints) and then modify the result to apply the preferred size
if defined.