Invalidate elements when they are moved.
Previously I had assumed that when an element was moved but not resized, it
need not be revalidated because everything would just render the same at the
different offset. But this is not accurate now that we round positions to
physical pixel coordinates.

Moving an element could change its rounding (which happens when it is
revalidate) and could potentially change the rounding of any of its children,
so it needs to go through a full revalidation.

Fixes: #88