XYFlicker: Fix acceleration vector direction/magnitude
Acceleration vector should have a direction opposite to the computed drag
velocity, but the previous implementation only computed the correct
acceleration vector for axis-parallel velocity vectors. For any other velocity
vector, the acceleration vector was parallel to the bisector of the quadrant
the velocity vector was in. This was making the scroll to
incorrectly stop on one axis before than the other for those cases.
On top of that the magnitude of the acceleration vector was up-scaled by a
factor of sqrt(2) for non axis-parallel velocity.

Fixed both the direction and the magnitude of the acceleration vector.
Additionally, Vector class is used instead of Point for representing
and manipulating velocity and acceleration, which seems more natural.