Fix glyph canvas layer not properly shrunk
The canvas layer held by a glyph instance is now properly shrunk when
the prepared size is smaller than the current one. This avoids strange
behaviours where a Button instance is re-laid out to a smaller size
but the inner canvas layer for the text is not resized and overflows
its parent layer.